refactor(theme): split up theme and make it ready for usage outside (#21)
