+++ .storybook/main.ts
@@ -12,6 +12,7 @@ const config: StorybookConfig = {
     '@storybook/addon-essentials',
     '@chromatic-com/storybook',
     '@storybook/addon-interactions',
+    'storybook-theme-css-vars',
   ],
   framework: {
     name: '@storybook/nextjs',
