+++ .storybook/preview.ts
@@ -8,6 +8,12 @@ const preview: Preview = {
         date: /Date$/i,
       },
     },
+    theme: {
+      selector: "body (or your selector with data attribute)",
+      dataAttr: "data-theme (or your data attribute)",
+      nameLightTheme: "light (or your name of light theme)",
+      nameDarkTheme: "dark (or your name of dark theme)",
+    },
   },
 };
