+++ src/components/organisms/Header.tsx
@@ -2,7 +2,7 @@ import { styled } from 'styled-components';
 import Link from 'next/link';
 import { FunctionComponent } from 'react';
 
-import { cssVars } from '@/styles/theme';
+import cssVars from '@/styles/css_vars';
 import Marquee from '@/components/molecules/Marquee';
 import MainNavigation from '@/components/organisms/MainNavigation';
 import UserNavigation from '@/components/organisms/UserNavigation';
