+++ src/components/templates/LockedContent.tsx
@@ -6,7 +6,7 @@ import React, {
 } from 'react';
 import { styled } from 'styled-components';
 
-import { cssVars } from '@/styles/theme';
+import cssVars from '@/styles/css_vars';
 
 import EscapingElement from './EscapingElement';
