+++ src/features/articles/components/CoverPlaceholder.tsx
@@ -1,17 +1,16 @@
 import { styled } from 'styled-components';
 import { FunctionComponent } from 'react';
 
-import { ThemeProps } from '@/styles/theme';
+import { ThemeColors } from '@/types';
 
 // SVG generated by https://pattern.monster/
 const Container = styled.div<{
   width: number;
   height: number;
-  theme: ThemeProps;
+  theme: ThemeColors;
 }>`
   position: relative;
-  background-image: ${({ theme }) =>
-    generateBackground(theme.colors.secondary)};
+  background-image: ${({ theme }) => generateBackground(theme.secondary)};
   background-repeat: repeat;
   padding-bottom: ${({ width, height }) => `${(height / width) * 100}%`};
