+++ src/providers/ThemeProvider.tsx
@@ -3,15 +3,13 @@ import { ThemeProvider as StyledComponentThemeProvider } from 'styled-components
 import { Provider as ReduxProvider } from 'react-redux';
 import { PersistGate } from 'reduxjs-toolkit-persist/integration/react';
 
-import {
-  DarkTheme,
-  DarkThemeStyle,
-  LightTheme,
-  LightThemeStyle,
-} from '@/styles/theme';
 import * as redux from '@/redux/store';
 import registerIcons from '@/utils/icons';
 import useUserColorScheme from '@/hooks/useUserColorScheme';
+import darkThemeColors from '@/styles/themes/dark_theme_colors';
+import lightThemeColors from '@/styles/themes/light_theme_colors';
+import DarkThemeVarStyle from '@/styles/themes/DarkThemeVarStyle';
+import LightThemeVarStyle from '@/styles/themes/LightThemeVarStyle';
 
 const ThemeProvider: FunctionComponent<PropsWithChildren> = ({ children }) => {
   const colorScheme = useUserColorScheme();
@@ -22,10 +20,10 @@ const ThemeProvider: FunctionComponent<PropsWithChildren> = ({ children }) => {
   return (
     <ReduxProvider store={redux.store}>
       <PersistGate loading={null} persistor={redux.persistor}>
-        {isDarkMode && <DarkThemeStyle />}
-        {!isDarkMode && <LightThemeStyle />}
+        {isDarkMode && <DarkThemeVarStyle />}
+        {!isDarkMode && <LightThemeVarStyle />}
         <StyledComponentThemeProvider
-          theme={isDarkMode ? DarkTheme : LightTheme}>
+          theme={isDarkMode ? darkThemeColors : lightThemeColors}>
           {children}
         </StyledComponentThemeProvider>
       </PersistGate>
