+++ src/styles/themes/dark_theme_colors.ts
@@ -1,0 +1,22 @@
+import { type ThemeColors } from '@/types';
+
+const darkThemeColors: ThemeColors = {
+  background: '#050416',
+  primary: '#00bcd4',
+  primaryAlt: '#00acc1',
+  secondary: '#ffc107',
+  secondaryAlt: '#ffb300',
+  tertiary: '#ff5722',
+  tertiaryAlt: '#ff5722',
+  surface: '#141221',
+  error: '#f44336',
+  onPrimary: '#ffffff',
+  onSecondary: '#000000',
+  onTertiary: '#000000',
+  onBackground: '#fff',
+  onSurface: '#fff',
+  onError: '#fff',
+  dimmer: '#464646b5',
+};
+
+export default darkThemeColors;
