+++ src/styles/themes/light_theme_colors.ts
@@ -1,0 +1,22 @@
+import { ThemeColors } from '@/types';
+
+const lightThemeColors: ThemeColors = {
+  background: '#fafafa',
+  primary: '#00b0c0',
+  primaryAlt: '#028aa5',
+  secondary: '#991199',
+  secondaryAlt: '#660159',
+  tertiary: '#ee4917',
+  tertiaryAlt: '#d33d0f',
+  surface: '#f3f5ec',
+  error: '#f44034',
+  onPrimary: '#fff',
+  onSecondary: '#050416',
+  onTertiary: '#050416',
+  onBackground: '#050416',
+  onSurface: '#050416',
+  onError: '#050416',
+  dimmer: '#3a3a3abc',
+};
+
+export default lightThemeColors;
